Move 2D-only editor plugins to their own folder
