+++ editor/plugins/2d/camera_2d_editor_plugin.cpp
@@ -30,10 +30,10 @@
 
 #include "camera_2d_editor_plugin.h"
 
-#include "canvas_item_editor_plugin.h"
 #include "core/config/project_settings.h"
 #include "editor/editor_node.h"
 #include "editor/editor_undo_redo_manager.h"
+#include "editor/plugins/canvas_item_editor_plugin.h"
 #include "editor/themes/editor_scale.h"
 #include "scene/2d/camera_2d.h"
 #include "scene/gui/label.h"
