+++ editor/plugins/2d/camera_2d_editor_plugin.h
