+++ editor/plugins/2d/cast_2d_editor_plugin.cpp
@@ -30,9 +30,9 @@
 
 #include "cast_2d_editor_plugin.h"
 
-#include "canvas_item_editor_plugin.h"
 #include "editor/editor_node.h"
 #include "editor/editor_undo_redo_manager.h"
+#include "editor/plugins/canvas_item_editor_plugin.h"
 #include "scene/2d/physics/ray_cast_2d.h"
 #include "scene/2d/physics/shape_cast_2d.h"
 #include "scene/main/viewport.h"
