+++ editor/plugins/2d/light_occluder_2d_editor_plugin.cpp
