+++ editor/plugins/2d/light_occluder_2d_editor_plugin.h
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/*  line_2d_editor_plugin.h                                               */
+/*  light_occluder_2d_editor_plugin.h                                     */
 /**************************************************************************/
 /*                         This file is part of:                          */
 /*                             GODOT ENGINE                               */
@@ -30,27 +30,34 @@
 
 #pragma once
 
-#include "editor/plugins/abstract_polygon_2d_editor.h"
-#include "scene/2d/line_2d.h"
+#include "editor/plugins/2d/abstract_polygon_2d_editor.h"
+#include "scene/2d/light_occluder_2d.h"
 
-class Line2DEditor : public AbstractPolygon2DEditor {
-	GDCLASS(Line2DEditor, AbstractPolygon2DEditor);
+class LightOccluder2DEditor : public AbstractPolygon2DEditor {
+	GDCLASS(LightOccluder2DEditor, AbstractPolygon2DEditor);
 
-	Line2D *node = nullptr;
+	LightOccluder2D *node = nullptr;
+
+	Ref<OccluderPolygon2D> _ensure_occluder() const;
 
 protected:
 	virtual Node2D *_get_node() const override;
-	virtual void _set_node(Node *p_line) override;
+	virtual void _set_node(Node *p_polygon) override;
 
 	virtual bool _is_line() const override;
+	virtual int _get_polygon_count() const override;
 	virtual Variant _get_polygon(int p_idx) const override;
 	virtual void _set_polygon(int p_idx, const Variant &p_polygon) const override;
+
 	virtual void _action_set_polygon(int p_idx, const Variant &p_previous, const Variant &p_polygon) override;
+
+	virtual bool _has_resource() const override;
+	virtual void _create_resource() override;
 };
 
-class Line2DEditorPlugin : public AbstractPolygon2DEditorPlugin {
-	GDCLASS(Line2DEditorPlugin, AbstractPolygon2DEditorPlugin);
+class LightOccluder2DEditorPlugin : public AbstractPolygon2DEditorPlugin {
+	GDCLASS(LightOccluder2DEditorPlugin, AbstractPolygon2DEditorPlugin);
 
 public:
-	Line2DEditorPlugin();
+	LightOccluder2DEditorPlugin();
 };
