+++ editor/plugins/2d/path_2d_editor_plugin.h
