+++ editor/plugins/2d/tiles/atlas_merging_dialog.h
