+++ editor/plugins/2d/tiles/tile_proxies_manager_dialog.h
