+++ editor/plugins/2d/tiles/tile_set_atlas_source_editor.h
