+++ editor/plugins/2d/tiles/tile_set_scenes_collection_source_editor.cpp
@@ -36,7 +36,7 @@
 #include "editor/editor_settings.h"
 #include "editor/editor_undo_redo_manager.h"
 #include "editor/gui/editor_file_dialog.h"
-#include "editor/plugins/tiles/tile_set_editor.h"
+#include "editor/plugins/2d/tiles/tile_set_editor.h"
 #include "editor/themes/editor_scale.h"
 
 #include "scene/gui/button.h"
