+++ modules/navigation_2d/editor/navigation_region_2d_editor_plugin.h
@@ -30,7 +30,8 @@
 
 #pragma once
 
-#include "editor/plugins/abstract_polygon_2d_editor.h"
+#include "editor/plugins/2d/abstract_polygon_2d_editor.h"
+
 #include "editor/plugins/editor_plugin.h"
 
 class AcceptDialog;
