+++ scu_builders.py
@@ -299,8 +299,9 @@ def generate_scu_files(max_includes_per_scu):
     process_folder(["editor/import"])
     process_folder(["editor/import/3d"])
     process_folder(["editor/plugins"])
+    process_folder(["editor/plugins/2d"])
+    process_folder(["editor/plugins/2d/tiles"])
     process_folder(["editor/plugins/gizmos"])
-    process_folder(["editor/plugins/tiles"])
 
     process_folder(["platform/android/export"])
     process_folder(["platform/ios/export"])
